Address review: home the lambda sentinel predicate in helpers/yaml
is_lambda_sentinel is a generic wire-format predicate, not automation-
specific. Move it from models.automations to helpers/yaml/scalar so the
generic component emitter no longer reaches into the automations model;
both consumers now depend the normal direction (controllers/helpers ->
helpers/yaml).
